// frontend/src/api/ApiUrl.ts
export enum ApiUrl {
  Subscriptions = '/api/subscriptions',
  Testimonials = '/api/testimonials',
  Products = '/api/products',
  LatestProducts = '/api/products/latest',
  Users = '/api/users',
  Auth = '/api/auth',
  Metadata = '/api/metadata',
  Goto = '/api/goto',
  Render = '/api/render',
  Spawn = '/api/spawn',
  File = '/api/file',
  NestedJson = '/api/nestedJson',
  Partners = '/api/partners',
  Email = '/api/email',
  Chat = '/api/chat',
  Missions = '/api/missions',
  App = '/api'
}